Seperated account's & account_extension's views
