+++ app/views/__init__.py
@@ -6,3 +6,4 @@ from .phone import phone_blueprint  # noqa
 from .reseller import reseller_blueprint  # noqa
 from .user import user_blueprint  # noqa
 from .reseller_product import reseller_product_blueprint  # noqa
+from .account_extension import account_extension_blueprint # noqa
