+++ app/views/account.py
@@ -4,7 +4,7 @@ from flask_login import login_required
 
 from app.models import Account, Product, Reseller, AccountExtension, AccountChanges, Phone
 from app.models import ResellerProduct
-from app.forms import AccountForm, AccountExtensionForm
+from app.forms import AccountForm
 from ..database import db
 from app.logger import log
 from app.ninja import NinjaInvoice
@@ -102,17 +102,19 @@ def add_ninja_invoice(account: Account): # noqa E999
             break
     else:
         # need a new invoice
-        current_invoice = NinjaInvoice.add(account.reseller.ninja_client_id, invoice_date)
-    current_invoice.add_item(
-        ninja_product_name(account.product.name, account.months),
-        account.name,
-        cost=reseller_product.init_price if reseller_product else 0)
+        current_invoice = NinjaInvoice.add(account.reseller.ninja_client_id, invoice_date)    
+    if current_invoice:
+        current_invoice.add_item(
+            ninja_product_name(account.product.name, account.months),
+            account.name,
+            cost=reseller_product.init_price if reseller_product else 0)
     if account.phone.name != "None":
         phone_name = f"Phone-{account.phone.name}"
-        current_invoice.add_item(
-            phone_name,
-            account.name,
-            cost=account.phone.price)
+        if current_invoice:
+            current_invoice.add_item(
+                phone_name,
+                account.name,
+                cost=account.phone.price)
 
 
 @account_blueprint.route("/account_save", methods=["POST"])
@@ -171,49 +173,6 @@ def save():
         flash('Form validation error', 'danger')
         log(log.ERROR, "Form validation error")
     return redirect(url_for('account.edit', id=form.id.data))
-
-
-@account_blueprint.route("/account_ext_add", methods=["POST"])
-@login_required
-def ext_save():
-    form = AccountExtensionForm(request.form)
-    if form.validate_on_submit():
-        account = Account.query.filter(Account.id == request.form['account_id']).first()
-        if form.id.data < 0:
-            account_ext = AccountExtension()
-            account_ext.account_id = account.id
-            account_ext.reseller_id = account.reseller_id
-            account_ext.months = account.months
-            account_ext.extension_date = account.activation_date
-            account_ext.save()
-            account.reseller_id = form.reseller_id.data
-            account.months = form.months.data
-            # Check that months must be in 1-12
-            if not 0 < account.months <= 12:
-                flash('Mohths must be in 1-12', 'danger')
-                return redirect(url_for('account.edit', id=account.id))
-            account.activation_date = form.extension_date.data
-            account.save()
-            # Register product in Invoice Ninja
-            add_ninja_invoice(account)
-        else:
-            account_ext = AccountExtension.query.filter(AccountExtension.id == form.id.data).first()
-            if request.form['action'] == 'delete':
-                account_ext.delete()
-            else:
-                account_ext.reseller_id = form.reseller_id.data
-                account_ext.months = form.months.data
-                account_ext.extension_date = form.extension_date.data
-                # Check that months must be in 1-12
-                if not 0 < account_ext.months <= 12:
-                    flash('Mohths must be in 1-12', 'danger')
-                    return redirect(url_for('account.edit', id=account.id))
-                account_ext.save()
-
-        return redirect(url_for('account.edit', id=request.form['account_id']))
-    else:
-        flash('Form validation error', 'danger')
-    return redirect(url_for('account.edit', id=request.form['account_id']))
 
 
 @account_blueprint.route("/account_delete", methods=["GET"])
